Split password reset functionality into PasswordsController
